121. Best Time to Buy and Sell Stock
